Moved the gpu images2neibs to the cuda folder.
